# apps/backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitVibe API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the FitVibe REST API. Generated from documentation anchors until automated zod-to-openapi wiring is complete.",
    "contact": {
      "name": "FitVibe Engineering",
      "url": "https://fitvibe.app",
      "email": "engineering@fitvibe.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.fitvibe.app/api/v1",
      "description": "Production"
    },
    {
      "url": "https://staging.fitvibe.app/api/v1",
      "description": "Staging"
    },
    {
      "url": "http://localhost:4100/api/v1",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/users/{alias}": {
      "get": {
        "summary": "Get public profile by alias",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "E.AUTH.INVALID_CREDENTIALS"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "PublicProfile": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "stats": {
            "type": "object",
            "properties": {
              "completedSessions": {
                "type": "integer"
              },
              "badges": {
                "type": "integer"
              }
            }
          }
        },
        "required": ["alias"]
      },
...
    }
  }
}